Separate out caml_startup-dependent code into `runtime` library
